feat: update configuration
This updates the firmware configuration to follow the changes
from https://github.com/nrfconnect/sdk-nrf/pull/5662

This also removes individual builds for LTE-m and NB-IoT
because the device supports both modes and can fall back
if the other is not available.

Eventually forcing a network mode will be a runtime
configuration.

BREAKING CHANGE: individual builds for LTE-m and NB-IoT removed